style: moved api creation to common-lib
