+++ admin-bot/src/main/kotlin/com/github/heheteam/adminbot/states/CourseInfoState.kt
@@ -1,10 +1,10 @@
 package com.github.heheteam.adminbot.states
 
-import com.github.heheteam.commonlib.api.AdminApi
-import com.github.heheteam.commonlib.CourseStatistics
 import com.github.heheteam.adminbot.Keyboards
 import com.github.heheteam.adminbot.formatters.CourseStatisticsFormatter
 import com.github.heheteam.commonlib.Course
+import com.github.heheteam.commonlib.CourseStatistics
+import com.github.heheteam.commonlib.api.AdminApi
 import com.github.heheteam.commonlib.state.BotState
 import com.github.heheteam.commonlib.util.waitDataCallbackQueryWithUser
 import dev.inmo.micro_utils.fsm.common.State
