+++ admin-bot/src/main/kotlin/com/github/heheteam/adminbot/states/QueryCourseForEditing.kt
@@ -1,7 +1,7 @@
 package com.github.heheteam.adminbot.states
 
-import com.github.heheteam.commonlib.api.AdminApi
 import com.github.heheteam.commonlib.Course
+import com.github.heheteam.commonlib.api.AdminApi
 import com.github.heheteam.commonlib.state.BotStateWithHandlers
 import com.github.heheteam.commonlib.util.Unhandled
 import com.github.heheteam.commonlib.util.UpdateHandlersController
@@ -20,9 +20,9 @@ data class QueryCourseForEditing(override val context: User) :
   private val sentMessages = mutableListOf<AccessibleMessage>()
 
   override suspend fun intro(
-      bot: BehaviourContext,
-      service: AdminApi,
-      updateHandlersController: UpdateHandlersController<() -> Unit, Course?, Any>,
+    bot: BehaviourContext,
+    service: AdminApi,
+    updateHandlersController: UpdateHandlersController<() -> Unit, Course?, Any>,
   ) {
     val courses = service.getCourses().map { it.value }
     val coursesPicker = createCoursePicker(courses)
