+++ admin-bot/src/main/kotlin/com/github/heheteam/adminbot/states/RemoveStudentState.kt
@@ -1,6 +1,5 @@
 package com.github.heheteam.adminbot.states
 
-import com.github.heheteam.commonlib.api.AdminApi
 import com.github.heheteam.adminbot.Dialogues.manyIdsAlreadyDoNotExistForStudentRemoving
 import com.github.heheteam.adminbot.Dialogues.manyIdsAreGoodForStudentRemoving
 import com.github.heheteam.adminbot.Dialogues.manyStudentIdsDoNotExist
@@ -9,6 +8,7 @@ import com.github.heheteam.adminbot.Dialogues.oneIdAlreadyDoesNotExistForStudent
 import com.github.heheteam.adminbot.Dialogues.oneIdIsGoodForStudentRemoving
 import com.github.heheteam.adminbot.Dialogues.oneStudentIdDoesNotExist
 import com.github.heheteam.commonlib.Course
+import com.github.heheteam.commonlib.api.AdminApi
 import com.github.heheteam.commonlib.interfaces.StudentId
 import com.github.heheteam.commonlib.state.BotStateWithHandlers
 import com.github.heheteam.commonlib.util.UpdateHandlersController
@@ -34,9 +34,9 @@ class RemoveStudentState(override val context: User, val course: Course, val cou
   }
 
   override suspend fun intro(
-      bot: BehaviourContext,
-      service: AdminApi,
-      updateHandlersController: UpdateHandlersController<() -> Unit, String, Any>,
+    bot: BehaviourContext,
+    service: AdminApi,
+    updateHandlersController: UpdateHandlersController<() -> Unit, String, Any>,
   ) {
     val message =
       bot.send(context) {
@@ -114,9 +114,9 @@ class RemoveStudentState(override val context: User, val course: Course, val cou
   }
 
   override suspend fun sendResponse(
-      bot: BehaviourContext,
-      service: AdminApi,
-      response: List<String>,
+    bot: BehaviourContext,
+    service: AdminApi,
+    response: List<String>,
   ) {
     sentMessages.forEach { bot.delete(it) }
     response.forEach { msg -> bot.send(context, msg) }
