+++ common-lib/src/main/kotlin/com/github/heheteam/commonlib/database/DatabaseStudentStorage.kt
@@ -3,12 +3,12 @@ package com.github.heheteam.commonlib.database
 import com.github.heheteam.commonlib.BindError
 import com.github.heheteam.commonlib.ResolveError
 import com.github.heheteam.commonlib.Student
+import com.github.heheteam.commonlib.database.table.ParentStudents
+import com.github.heheteam.commonlib.database.table.StudentTable
 import com.github.heheteam.commonlib.interfaces.ParentId
 import com.github.heheteam.commonlib.interfaces.StudentId
 import com.github.heheteam.commonlib.interfaces.StudentStorage
 import com.github.heheteam.commonlib.interfaces.toStudentId
-import com.github.heheteam.commonlib.database.table.ParentStudents
-import com.github.heheteam.commonlib.database.table.StudentTable
 import com.github.michaelbull.result.Err
 import com.github.michaelbull.result.Ok
 import com.github.michaelbull.result.Result
