+++ common-lib/src/main/kotlin/com/github/heheteam/commonlib/decorators/GradeTableDecorator.kt
@@ -3,6 +3,7 @@ package com.github.heheteam.commonlib.decorators
 import com.github.heheteam.commonlib.Grade
 import com.github.heheteam.commonlib.Problem
 import com.github.heheteam.commonlib.SolutionAssessment
+import com.github.heheteam.commonlib.googlesheets.GoogleSheetsRatingRecorder
 import com.github.heheteam.commonlib.interfaces.AssignmentId
 import com.github.heheteam.commonlib.interfaces.CourseId
 import com.github.heheteam.commonlib.interfaces.GradeTable
@@ -12,7 +13,6 @@ import com.github.heheteam.commonlib.interfaces.ProblemId
 import com.github.heheteam.commonlib.interfaces.SolutionId
 import com.github.heheteam.commonlib.interfaces.StudentId
 import com.github.heheteam.commonlib.interfaces.TeacherId
-import com.github.heheteam.commonlib.googlesheets.GoogleSheetsRatingRecorder
 
 class GradeTableDecorator(
   private val gradeTable: GradeTable,
