+++ common-lib/src/main/kotlin/com/github/heheteam/commonlib/util/SampleFill.kt
@@ -1,6 +1,7 @@
 package com.github.heheteam.commonlib.util
 
 import com.github.heheteam.commonlib.ProblemDescription
+import com.github.heheteam.commonlib.database.reset
 import com.github.heheteam.commonlib.interfaces.AssignmentStorage
 import com.github.heheteam.commonlib.interfaces.CourseId
 import com.github.heheteam.commonlib.interfaces.CoursesDistributor
@@ -9,7 +10,6 @@ import com.github.heheteam.commonlib.interfaces.StudentId
 import com.github.heheteam.commonlib.interfaces.StudentStorage
 import com.github.heheteam.commonlib.interfaces.TeacherId
 import com.github.heheteam.commonlib.interfaces.TeacherStorage
-import com.github.heheteam.commonlib.database.reset
 import kotlin.time.Duration.Companion.minutes
 import kotlinx.datetime.Clock
 import kotlinx.datetime.TimeZone
