+++ student-bot/src/main/kotlin/com/github/heheteam/studentbot/StudentRun.kt
@@ -29,9 +29,9 @@ import kotlinx.coroutines.Dispatchers
 
 @OptIn(RiskFeature::class)
 suspend fun studentRun(
-    botToken: String,
-    studentApi: StudentApi,
-    developerOptions: DeveloperOptions? = DeveloperOptions(),
+  botToken: String,
+  studentApi: StudentApi,
+  developerOptions: DeveloperOptions? = DeveloperOptions(),
 ) =
   telegramBotWithBehaviourAndFSMAndStartLongPolling(
       botToken,
@@ -61,8 +61,8 @@ private fun reportExceptionAndPreserveState(state: State, e: Throwable): State {
 }
 
 private fun DefaultBehaviourContextWithFSM<State>.registerStates(
-    studentApi: StudentApi,
-    botToken: String,
+  studentApi: StudentApi,
+  botToken: String,
 ) {
   registerState<StartState, StudentApi>(studentApi)
   registerState<DeveloperStartState, StudentApi>(studentApi)
@@ -79,8 +79,8 @@ private fun DefaultBehaviourContextWithFSM<State>.registerStates(
 }
 
 private fun DefaultBehaviourContextWithFSM<State>.registerSendSolutionState(
-    botToken: String,
-    studentApi: StudentApi,
+  botToken: String,
+  studentApi: StudentApi,
 ) {
   strictlyOn<SendSolutionState> { state ->
     state.studentBotToken = botToken
