+++ student-bot/src/main/kotlin/com/github/heheteam/studentbot/state/MenuState.kt
@@ -1,5 +1,6 @@
 package com.github.heheteam.studentbot.state
 
+import com.github.heheteam.commonlib.api.StudentApi
 import com.github.heheteam.commonlib.interfaces.StudentId
 import com.github.heheteam.commonlib.interfaces.toStudentId
 import com.github.heheteam.commonlib.state.BotStateWithHandlers
@@ -16,7 +17,6 @@ import com.github.heheteam.studentbot.Keyboards
 import com.github.heheteam.studentbot.Keyboards.CHECK_DEADLINES
 import com.github.heheteam.studentbot.Keyboards.CHECK_GRADES
 import com.github.heheteam.studentbot.Keyboards.SEND_SOLUTION
-import com.github.heheteam.commonlib.api.StudentApi
 import dev.inmo.kslog.common.error
 import dev.inmo.kslog.common.logger
 import dev.inmo.micro_utils.fsm.common.State
