+++ student-bot/src/main/kotlin/com/github/heheteam/studentbot/state/StartState.kt
@@ -1,10 +1,10 @@
 package com.github.heheteam.studentbot.state
 
+import com.github.heheteam.commonlib.api.StudentApi
 import com.github.heheteam.commonlib.interfaces.StudentId
 import com.github.heheteam.commonlib.state.BotState
 import com.github.heheteam.commonlib.util.waitTextMessageWithUser
 import com.github.heheteam.studentbot.Dialogues
-import com.github.heheteam.commonlib.api.StudentApi
 import com.github.michaelbull.result.get
 import dev.inmo.micro_utils.fsm.common.State
 import dev.inmo.tgbotapi.extensions.api.send.media.sendSticker
