+++ teacher-bot/src/main/kotlin/com/github/heheteam/teacherbot/states/GradingMessagesFunctions.kt
@@ -2,9 +2,9 @@ package com.github.heheteam.teacherbot.states
 
 import com.github.heheteam.commonlib.Grade
 import com.github.heheteam.commonlib.SolutionAssessment
+import com.github.heheteam.commonlib.api.TeacherApi
 import com.github.heheteam.commonlib.interfaces.SolutionId
 import com.github.heheteam.commonlib.interfaces.TeacherId
-import com.github.heheteam.commonlib.api.TeacherApi
 import com.github.michaelbull.result.Err
 import com.github.michaelbull.result.Ok
 import com.github.michaelbull.result.Result
