+++ teacher-bot/src/main/kotlin/com/github/heheteam/teacherbot/states/StartState.kt
@@ -1,10 +1,10 @@
 package com.github.heheteam.teacherbot.states
 
+import com.github.heheteam.commonlib.api.TeacherApi
 import com.github.heheteam.commonlib.interfaces.TeacherId
 import com.github.heheteam.commonlib.state.BotState
 import com.github.heheteam.commonlib.util.waitTextMessageWithUser
 import com.github.heheteam.teacherbot.Dialogues
-import com.github.heheteam.commonlib.api.TeacherApi
 import com.github.michaelbull.result.get
 import dev.inmo.micro_utils.fsm.common.State
 import dev.inmo.tgbotapi.extensions.api.send.media.sendSticker
